Restructure project. Implement Router.
